Agregando funciones en demas tablas y carpetas dto y mapper

+++ Zorriana-Back/src/main/java/com/zorriana/backend/articulos/service/impl/ArticulosServiceImpl.java
@@ -4,6 +4,8 @@ import com.zorriana.backend.articulos.entity.Articulos;
 import com.zorriana.backend.articulos.repository.IArticulosRepository;
 import com.zorriana.backend.articulos.service.IArticulosService;
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,12 @@ public class ArticulosServiceImpl implements IArticulosService {
   }
 
   @Override
+  public Optional<Articulos> findById(Long id) {
+    return repository.findById(id);
+  }
+
+
+  @Override
   public void save(Articulos articulo) {
     repository.save(articulo);
   }
@@ -30,6 +38,9 @@ public class ArticulosServiceImpl implements IArticulosService {
 
   @Override
   public void deleteById(Long id) {
-    repository.deleteById(id);
+    Optional<Articulos> artOp = repository.findById(id);
+    if (artOp.isPresent()) {
+      repository.deleteById(id);
+    }
   }
 }
